doc: Change blog post links to a top-level placeholder
I'd like the upcoming blog post to point to items in the source
repo at a tag, should they ever move in the future. However, docs
rendered at this tag would point to 404'd links that are currently
just placeholders. (It's a chicken-or-egg.)

They now just point to the top-level research.nccgroup.com page,
and I will update them accordingly once the blog post is live.