python: Reworked payload_base and payload_offset code
This is around of touch-ups preceding changes aimed to fail
into a "don't deploy payloads" mode of operation.

Along the way there were some small fixes and improvements
that will likely yield some new behavior.

For example - allowing a kwargs-override of the payload_base pulled from
an existing JSON. (This previously resulted in an exception.)