Change README.md to a symlink
By creating a repository containing multiple components, one of which
is a Python package, I seem to have shot myself in the foot. Figures.

(Yes, git subrepos would have been an appropriate technical solution,
but this was not an option to me at the time.)

The problem is as follows:

We want the package hosted on PyPI to have a "long description" that
always matches the Depthcharge project's top-level README.md.
Previously, I had simply been loading this via a relative path (to the
parent directory) from within setup.py.

This was all well and good until I discovered that this effectively
broke document generation with ReadTheDocs.io; the doc build operation
is staged in a working directory assuming that setup.py resides in the
root; attempting to access ../README.md does not point where we expect.

I do not want to maintain two copies of the readme, nor remember to
run long_desc.py each time the README.md is touched. This changeset
shimmies things around with symlinks, which is Gross But Sufficient.

There are bigger fish to try. Calling this good enough to close #14.
We can address the next problem this causes when it rears its ugly head.