python: Make setup.py version check regex a superset of PEP440 reqt
[PEP440] requires that versions be in the form:

    [N!]N(.N)*[{a|b|rc}N][.postN][.devN]

Local version identifiers are also supported:

    <public version identifier>[+<local version label>]

My use of dashes violates this. However, I'd prefer that the setup.py
still work in light of such mistakes. The updated regex supports
".", "-", and "+".  The match for both .postN.devN together is also
included.

Ensuring strict compliance to the versioning that will be enforced
by PyPi is something should be enforced via unit tests or CI. Opened
issue #73 to track  this.

[PEP440]: https://www.python.org/dev/peps/pep-0440/#public-version-identifiers