payloads: Remove return_register
This payload is usually just used to retrieve a gd pointer. The
same can be done with return_memory_word (with no argument).

Given the lack of any significant use case (that can't be addressed
by a quick custom user payload) and the nuance associated with only
a subset of registers not being tainted by the payload execution
itself, I'm removing this.