depthcharge/arch: Rename is_aligned to is_word_aligned
In most cases, we want to check data structure elements for
word-alignment. The is_aligned() method has been renamed to
more explicitly describe the alignment check.

However, we also may want to check that an unaligned memory access is
permissible before attempting it. is_allowed_access() may be used for
this.