depthcharge: Use 0 for CpCrashRegisterReader src addr on AARCH64
It appears that having src == dest yields no effect, even for
a bogus address. After further digging, this isn't unique to
AARCH64, but rather, is no longer working on ARM either.

It appears that a U-Boot change to use memcpy() was made prior to the
2017.01 release (c2538421b28424b9705865e838c5fba19c9dc651). The
cp-based data abort appears to work fine fine through 2016.11.

The special AARCH64 case added in this commit appears to be sufficient
to induce a data abort on a modern U-Boot.

We'll have to tackle the ARM side of things later with a special case
and/or version check. It seems possible to cause a crash still on
newer U-Boot versions, but not in a generic way.

For example, on an i.MX6ULL, setting the source address to an unaligned
address in a reserved region seems to do the trick:

    In:    serial@2020000
    Out:   serial@2020000
    Err:   serial@2020000
    Net:   eth0: ethernet@2188000
    Hit any key to stop autoboot:  0

    IMX6ULL-SOM U-Boot > cp.l 0x20000001 $loadaddr 1
    data abort
    pc : [<9ffa48ca>]          lr : [<9ffaac9d>]
    reloc pc : [<878008ca>]    lr : [<87806c9d>]
    sp : 9ef8d36c  ip : 00000001     fp : 00000004
    r10: 00000000  r9 : 9ef93ee0     r8 : 00000004
    r7 : 00000000  r6 : 20000001     r5 : 82000000  r4 : 9efa3668
    r3 : 00000000  r2 : 00000000     r1 : 20000000  r0 : 82000000
    Flags: Nzcv  IRQs off  FIQs off  Mode SVC_32 (T)
    Code: f021 0103 f1bc 0f02 (f851) eb04
    Resetting CPU ...